Removed old console.warns and fixed bugs related to the refactor
