+++ client/src/components/Container/ItemSelector/styles.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  width: 100%;
+`;
